[#43] refactor: 컴포넌트는 components 폴더 내부로 이동
