+++ src/app/place/page.tsx
@@ -1,4 +1,4 @@
-import Place from "@/app/place/Place";
+import Place from "@/components/place/Place";
 
 export default function Page() {
   return <Place />;
