+++ src/app/signUp/page.tsx
@@ -1,4 +1,4 @@
-import SignUp from "@/app/signUp/SignUp";
+import SignUp from "@/components/auth/SignUp";
 
 export default function Page() {
   return <SignUp />;
