+++ src/components/place/SearchedPlace.tsx
