+++ src/components/record/HydratedRecordDetail.tsx
@@ -1,6 +1,6 @@
 import { dehydrate } from "@tanstack/query-core";
 import { HydrationBoundary } from "@tanstack/react-query";
-import RecordDetail from "@/app/record/[recordId]/RecordDetail";
+import RecordDetail from "@/components/record/RecordDetail";
 import getQueryClient from "@/hooks/getQueryClient";
 
 export default async function HydratedRecordDetail() {
