+++ src/components/record/RecordDetail.tsx
